+++ cypress/support/e2e.js
@@ -4,6 +4,7 @@
  */
 
 import '../utils/commands';
+import '../utils/commands.osd';
 import '../utils/apps/commands';
 import '../utils/dashboards/workspace-plugin/commands';
 import '../utils/dashboards/commands';
